Fix product variation reload issue
When selecting a variation on the product page, the "Add to cart" button doesn't display the loading symbol (eg. https://bravetheme.neto.com.au/brave-basic-tee). So if you quickly change the variation and click ATC it will add the previously selected product to cart if you don't give it enough time to load.